separate cluster and path utilities
