+++ meddlr/utils/__init__.py
@@ -1,1 +1,2 @@
 from . import cluster  # noqa: F401
+from . import path  # noqa: F401
